+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/client/TigerRemoteProxyClientTest.java
@@ -641,6 +641,7 @@ class TigerRemoteProxyClientTest {
           .sender(RbelHostname.fromString("server:80").get())
           .receiver(RbelHostname.fromString("client:80").get());
     }
+    builder.request(request);
     return builder.build();
   }